Submitted proof paper to Taylor & Francis
I was not sure where to submit the proof paper when I was writing it,
but I ended up deciding on Taylor & Francis. Unfortunately, I used the
elsarticle latex documentclass to write the paper, so the footer on the
title page said "Preprint submitted to Elsevier". This commit changes
Elsevier to Taylor & Francis in this footer and elsewhere.